+++ framework/domain/src/main/java/is/codion/framework/domain/entity/condition/Condition.java
@@ -124,7 +124,7 @@ public interface Condition {
 			return compositeEqualCondition(columnMap, EQUAL, valueMap);
 		}
 
-		return key.column().equalTo(key.get());
+		return key.column().equalTo(key.value());
 	}
 
 	/**